Move autograd function for CrossMapLRN2d from being backend specific to modules/_functions.
This is to get rid of backend-specific dispatch in modules; this autograd function is no longer backend specific so
doesn't need to be in a backend specific location.
