+++ torch/nn/_functions/thnn/__init__.py
@@ -1,3 +1,1 @@
 _all_functions = []
-
-from .normalization import *  # noqa: F401
