+++ torch/nn/modules/normalization.py
@@ -2,6 +2,7 @@ import torch
 import numbers
 from torch.nn.parameter import Parameter
 from .module import Module
+from ._functions import CrossMapLRN2d as _cross_map_lrn2d
 from .. import functional as F
 from .. import init
 
@@ -61,8 +62,8 @@ class CrossMapLRN2d(Module):
         self.k = k
 
     def forward(self, input):
-        return self._backend.CrossMapLRN2d.apply(input, self.size, self.alpha, self.beta,
-                                                 self.k)
+        return _cross_map_lrn2d.apply(input, self.size, self.alpha, self.beta,
+                                      self.k)
 
     def extra_repr(self):
         return '{size}, alpha={alpha}, beta={beta}, k={k}'.format(**self.__dict__)
